requirements.txt
Django
django-bootstrap-modal-forms
numpy
pandas
djangorestframework